refactor: split machine migration testing into multiple files
